fix: Fixed the PriorityRankScore calculation logics
Re-updated the PriorityRankScore(PRS) with missing logics in
calculation, in which the taskCapacity value is not correctly
calculated.
